Reconfigured the directory structure of the Cybersprint project to enhance organization and maintainability. The changes include: - Moved Flashcard components to a dedicated `Flashcard` directory. - Moved Glossary components to a dedicated `Glossary` directory. - Moved Homepage components to a dedicated `Homepage` directory. - Moved Navbar and Footer components to a dedicated `Navbar-Footer` directory.
