+++ cybersprint/src/App.jsx
@@ -1,20 +1,17 @@
 import { Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Dashboard from './pages/Dashboard';
-import Navbar from './components/Navbar';
+import Navbar from "./components/Navbar-Footer/Navbar";
 
 export default function App() {
   return (
     <>
-      {/* Need to include Navbar here */}
+      {/* Navbar appears on all routes */}
       <Navbar />
 
-      {/* Define routes using React Router */}
+      {/* Define routes */}
       <Routes>
-        {/* When user visits "/", they see Home.jsx */}
         <Route path="/" element={<Home />} />
-
-        {/* When user visits "/dashboard", they see Dashboard.jsx */}
         <Route path="/dashboard" element={<Dashboard />} />
       </Routes>
     </>
