+++ cybersprint/src/components/Glossary/Glossary.jsx
@@ -1,26 +1,29 @@
 import { useState } from "react";
-import { glossaryTerms } from "../data/glossaryData";
+import { glossaryTerms } from "../../data/glossaryData";
 import GlossaryCard from "./GlossaryCard";
-import '../styles/Glossary.css';
+import "../../styles/Glossary/Glossary.css";
 
 const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ".split("");
 
 export default function Glossary() {
   const [activeLetter, setActiveLetter] = useState("A");
-  const filtered = glossaryTerms.filter(term => term.letter === activeLetter);
+
+  // Filter terms based on selected letter
+  const filteredTerms = glossaryTerms.filter(term => term.letter === activeLetter);
 
   return (
     <section className="glossary">
       <div className="glossary-wrapper">
         <h2 className="glossary-heading">Cyber A–Z Vocabulary</h2>
 
-        {/* Alphabet Grid */}
+        {/* Alphabet Navigation */}
         <div className="glossary-letters">
-          {alphabet.map((letter) => (
+          {alphabet.map(letter => (
             <button
               key={letter}
-              className={`glossary-letter ${activeLetter === letter ? 'active' : ''}`}
+              className={`glossary-letter ${activeLetter === letter ? "active" : ""}`}
               onClick={() => setActiveLetter(letter)}
+              aria-label={`Show terms that start with ${letter}`}
             >
               {letter}
             </button>
@@ -29,8 +32,8 @@ export default function Glossary() {
 
         {/* Term Cards */}
         <div className="glossary-cards">
-          {filtered.length > 0 ? (
-            filtered.map((item, idx) => (
+          {filteredTerms.length > 0 ? (
+            filteredTerms.map((item, idx) => (
               <GlossaryCard key={idx} {...item} />
             ))
           ) : (
