+++ cybersprint/src/components/Navbar-Footer/Footer.jsx
@@ -1,3 +1,5 @@
+import '../../styles/Navbar-Footer/Footer.css';
+
 export default function Footer() {
   return (
 <footer className="footer">
