+++ cybersprint/src/index.css
@@ -31,7 +31,6 @@ body {
 * {
   transition: all 0.2s ease-in-out;
 }
-
 /* Scrollbar: subtle for dark mode (optional) */
 ::-webkit-scrollbar {
   width: 8px;
