--- cybersprint/src/styles/Dashboard.css
@@ -1,75 +1,0 @@
-/* Background + text global */
-.dashboard {
-  background-color: #0F0F1A;
-  color: white;
-  font-family: 'Poppins', sans-serif;
-  min-height: 100vh;
-}
-
-/* Main container */
-.dashboard-container {
-  max-width: 96rem;
-  margin: 0 auto;
-  padding: 3rem 1.5rem;
-}
-
-/* Section titles */
-.dashboard-title {
-  font-size: 1.75rem;
-  font-weight: 800;
-  letter-spacing: -0.02em;
-  margin-bottom: 2rem;
-}
-
-/* Card blocks */
-.dashboard-card {
-  background-color: #1A1A2C;
-  border-radius: 1rem;
-  padding: 1rem;
-  text-align: center;
-  box-shadow: 0 6px 16px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-}
-
-.dashboard-card:hover {
-  transform: translateY(-4px);
-  box-shadow: 0 10px 24px rgba(124, 58, 237, 0.2);
-}
-
-.dashboard-card .value {
-  font-size: 2rem;
-  font-weight: bold;
-  color: #a855f7;
-}
-
-.dashboard-card .label {
-  font-size: 0.875rem;
-  color: #ccc;
-  margin-top: 0.25rem;
-}
-
-/* Grid layout sections */
-.dashboard-grid {
-  display: grid;
-  gap: 2.5rem;
-}
-
-.dashboard-grid-4 {
-  grid-template-columns: repeat(1, 1fr);
-}
-
-@media (min-width: 640px) {
-  .dashboard-grid-4 {
-    grid-template-columns: repeat(2, 1fr);
-  }
-}
-
-@media (min-width: 1024px) {
-  .dashboard-grid-4 {
-    grid-template-columns: repeat(4, 1fr);
-  }
-
-  .dashboard-grid-2fr-1fr {
-    grid-template-columns: 2fr 1fr;
-  }
-}
