+++ cybersprint/src/styles/Flashcard/FlashcardItem.css
