+++ cybersprint/src/styles/Glossary/Glossary.css
@@ -98,8 +98,6 @@
     padding: 0.35rem 0.75rem;
     border-radius: 9999px;
   }
-  
-  /* ... existing styles remain unchanged ... */
 
 /* === Additional Category Colors === */
 .category-core {
