+++ cybersprint/src/styles/Homepage/Dashboard.css
@@ -1,0 +1,105 @@
+/* ================================
+   🌌 Dashboard Page Global Styles
+   Affects the entire dashboard UI layout, background, fonts, and colors.
+   ================================ */
+
+/* Main page container - sets background, text color, and font */
+.dashboard {
+  background-color: #0F0F1A; /* Dark background for cyber theme */
+  color: white;              /* Light text for contrast */
+  font-family: 'Poppins', sans-serif; /* Consistent brand font */
+  min-height: 100vh;         /* Full viewport height */
+  padding: 3rem 1.5rem;      /* Page padding */
+}
+
+/* Central content wrapper - limits width and centers content */
+.dashboard-container {
+  max-width: 96rem;          /* Prevents overly wide content */
+  margin: 0 auto;            /* Centers the container horizontally */
+}
+
+/* ================================
+   🧭 Section Headings
+   Used for titles like "Your Progress", "Next Module", etc.
+   ================================ */
+.dashboard-title {
+  font-size: 1.75rem;        /* Large, noticeable title */
+  font-weight: 800;          /* Bold for emphasis */
+  letter-spacing: -0.02em;   /* Slight tightening of text */
+  margin-bottom: 2rem;       /* Space below the title */
+}
+
+/* ================================
+   🧱 Grid Layouts
+   Manages how dashboard cards/modules/quizzes appear in grid
+   ================================ */
+
+/* Base grid container for content sections */
+.dashboard-grid {
+  display: grid;             /* Enables grid layout */
+  gap: 2rem;                 /* Space between grid items */
+}
+
+/* Grid with 1 column on mobile, expands to 2 and 4 columns */
+.dashboard-grid-4 {
+  grid-template-columns: 1fr; /* One column by default */
+}
+
+@media (min-width: 640px) {
+  .dashboard-grid-4 {
+    grid-template-columns: repeat(2, 1fr); /* Two columns on small screens */
+  }
+}
+
+@media (min-width: 1024px) {
+  .dashboard-grid-4 {
+    grid-template-columns: repeat(4, 1fr); /* Four columns on large screens */
+  }
+}
+
+/* Optional layout for split sections like a 2/3 content + 1/3 sidebar */
+.dashboard-grid-2fr-1fr {
+  display: grid;
+  gap: 2rem;
+}
+
+@media (min-width: 1024px) {
+  .dashboard-grid-2fr-1fr {
+    grid-template-columns: 2fr 1fr; /* Wider left column, narrow right */
+  }
+}
+
+/* ================================
+   🔲 Dashboard Cards
+   Affects individual cards for stats, modules, progress, etc.
+   ================================ */
+.dashboard-card {
+  background-color: #1A1A2C; /* Slightly lighter than background */
+  border-radius: 1rem;       /* Rounded edges for soft look */
+  padding: 1.5rem;           /* Internal padding */
+  text-align: center;        /* Centers content inside */
+  box-shadow: 0 6px 16px rgba(0, 0, 0, 0.1); /* Soft shadow */
+  transition: transform 0.3s ease, box-shadow 0.3s ease; /* Hover effects */
+}
+
+/* On hover: elevate the card slightly */
+.dashboard-card:hover {
+  transform: translateY(-4px); /* Moves card up slightly */
+  box-shadow: 0 10px 24px rgba(124, 58, 237, 0.2); /* Adds vibrant glow */
+}
+
+/* ================================
+   📊 Card Content Styling
+   Used inside each card for value + label text
+   ================================ */
+.dashboard-card .value {
+  font-size: 2rem;           /* Large number (e.g. "85%") */
+  font-weight: bold;
+  color: #a855f7;            /* Highlighted purple color */
+}
+
+.dashboard-card .label {
+  font-size: 0.875rem;       /* Smaller label text below value */
+  color: #ccc;               /* Light gray text */
+  margin-top: 0.5rem;
+}
