+++ cybersprint/src/styles/Homepage/HeroSection.css
@@ -1,0 +1,88 @@
+/* Background and container */
+.hero-section {
+  position: relative;
+  background: radial-gradient(circle at 30% 30%, #1A1A2C, #0F0F1A);
+  min-height: 80vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0 1rem;
+  overflow: hidden;
+}
+
+.hero-section::before {
+  content: '';
+  position: absolute;
+  top: -100px;
+  left: 50%;
+  transform: translateX(-50%);
+  width: 600px;
+  height: 600px;
+  background: radial-gradient(circle, rgba(124, 58, 237, 0.15), transparent);
+  z-index: 0;
+  border-radius: 50%;
+  pointer-events: none;
+  animation: pulseGlow 6s ease-in-out infinite;
+}
+
+/* Hero content */
+.hero-content {
+  max-width: 720px;
+  text-align: center;
+  z-index: 1;
+}
+
+/* Headline */
+.hero-heading {
+  font-size: 2.5rem;
+  font-weight: 800;
+  color: white;
+  line-height: 1.2;
+  margin-bottom: 1.5rem;
+  text-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
+}
+
+@media (min-width: 640px) {
+  .hero-heading {
+    font-size: 3.75rem;
+  }
+}
+
+/* Subheading */
+.hero-subheading {
+  font-size: 1.125rem;
+  color: #d1d5db;
+  margin-bottom: 2.5rem;
+}
+
+/* Call-to-action button */
+.start-button-blend {
+  padding: 0.75rem 2rem;
+  font-size: 0.9rem;
+  font-weight: 600;
+  border-radius: 9999px;
+  background: white;
+  color: #0F0F1A;
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.15);
+  transition: all 0.3s ease;
+  border: none;
+  cursor: pointer;
+}
+
+.start-button-blend:hover {
+  background: #f1f1f1;
+  transform: translateY(-2px);
+  box-shadow: 0 15px 25px rgba(124, 58, 237, 0.2);
+}
+
+/* Glow animation */
+@keyframes pulseGlow {
+  0%, 100% {
+    opacity: 0.3;
+    transform: translateX(-50%) scale(1);
+  }
+  50% {
+    opacity: 0.5;
+    transform: translateX(-50%) scale(1.05);
+  }
+}
